refactor path calculation
