+++ gradle.properties
@@ -1,5 +1,5 @@
 # Kompendium
-project.version=1.7.0
+project.version=1.8.0
 # Kotlin
 kotlin.code.style=official
 # Gradle
