+++ kompendium-core/src/main/kotlin/io/bkbn/kompendium/Kompendium.kt
@@ -5,7 +5,7 @@ import io.bkbn.kompendium.models.meta.SchemaMap
 import io.bkbn.kompendium.models.oas.OpenApiSpec
 import io.bkbn.kompendium.models.oas.OpenApiSpecInfo
 import io.bkbn.kompendium.models.oas.TypedSchema
-import io.bkbn.kompendium.path.CorePathCalculator
+import io.bkbn.kompendium.path.IPathCalculator
 import io.bkbn.kompendium.path.PathCalculator
 import io.ktor.routing.Route
 import io.ktor.routing.RouteSelector
@@ -16,8 +16,6 @@ import kotlin.reflect.KClass
  */
 object Kompendium {
 
-  private val pathHandler: MutableMap<RouteSelector, (Route?, String) -> String> = mutableMapOf()
-
   var errorMap: ErrorMap = emptyMap()
   var cache: SchemaMap = emptyMap()
 
@@ -27,7 +25,7 @@ object Kompendium {
     paths = mutableMapOf()
   )
 
-  var pathCalculator: PathCalculator = CorePathCalculator()
+  fun calculate(route: Route) = PathCalculator.calculate(route)
 
   fun resetSchema() {
     openApiSpec = OpenApiSpec(
@@ -42,7 +40,11 @@ object Kompendium {
     cache = cache.plus(clazz.simpleName!! to schema)
   }
 
-  fun addCustomRouteHandler(selector: RouteSelector, handler: (Route?, String) -> String) {
-    pathHandler[selector] = handler
+  fun <T : RouteSelector> addCustomRouteHandler(
+    selector: KClass<T>,
+    handler: IPathCalculator.(Route, String) -> String
+  ) {
+    PathCalculator.addCustomRouteHandler(selector, handler)
   }
+
 }
