+++ kompendium-core/src/main/kotlin/io/bkbn/kompendium/Notarized.kt
@@ -16,6 +16,7 @@ import io.bkbn.kompendium.models.meta.MethodInfo.PutInfo
 import io.bkbn.kompendium.models.meta.MethodInfo.DeleteInfo
 import io.bkbn.kompendium.models.meta.ResponseInfo
 import io.bkbn.kompendium.models.oas.OpenApiSpecPathItem
+import io.bkbn.kompendium.path.PathCalculator
 
 /**
  * Notarization methods are the primary way that a Ktor API using Kompendium differentiates
@@ -36,7 +37,7 @@ object Notarized {
     noinline body: PipelineInterceptor<Unit, ApplicationCall>
   ): Route =
     KompendiumPreFlight.methodNotarizationPreFlight<TParam, Unit, TResp>() { paramType, requestType, responseType ->
-      val path = Kompendium.pathCalculator.calculate(this)
+      val path = Kompendium.calculate(this)
       Kompendium.openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
       Kompendium.openApiSpec.paths[path]?.get = parseMethodInfo(info, paramType, requestType, responseType)
       return method(HttpMethod.Get) { handle(body) }
@@ -55,7 +56,7 @@ object Notarized {
     noinline body: PipelineInterceptor<Unit, ApplicationCall>
   ): Route =
     KompendiumPreFlight.methodNotarizationPreFlight<TParam, TReq, TResp>() { paramType, requestType, responseType ->
-      val path = Kompendium.pathCalculator.calculate(this)
+      val path = Kompendium.calculate(this)
       Kompendium.openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
       Kompendium.openApiSpec.paths[path]?.post = parseMethodInfo(info, paramType, requestType, responseType)
       return method(HttpMethod.Post) { handle(body) }
@@ -74,7 +75,7 @@ object Notarized {
     noinline body: PipelineInterceptor<Unit, ApplicationCall>,
   ): Route =
     KompendiumPreFlight.methodNotarizationPreFlight<TParam, TReq, TResp>() { paramType, requestType, responseType ->
-      val path = Kompendium.pathCalculator.calculate(this)
+      val path = Kompendium.calculate(this)
       Kompendium.openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
       Kompendium.openApiSpec.paths[path]?.put =
         parseMethodInfo(info, paramType, requestType, responseType)
@@ -93,7 +94,7 @@ object Notarized {
     noinline body: PipelineInterceptor<Unit, ApplicationCall>
   ): Route =
     KompendiumPreFlight.methodNotarizationPreFlight<TParam, Unit, TResp> { paramType, requestType, responseType ->
-      val path = Kompendium.pathCalculator.calculate(this)
+      val path = Kompendium.calculate(this)
       Kompendium.openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
       Kompendium.openApiSpec.paths[path]?.delete = parseMethodInfo(info, paramType, requestType, responseType)
       return method(HttpMethod.Delete) { handle(body) }
